+++ docs/history.rst
@@ -7,8 +7,12 @@ History
 Version 0.5.2 - Unreleased
 --------------------------
 
-- [ ] [collect] Dumped cover filepath is incorrect #4;
-- [ ] [collect] Directory checksum should not use computed cover filename with UUID #5;
+* Fixed dumped cover item filepath to be relative to the dump file;
+* Fixed directory checksum to be done with cover source path, so no UUID is involved
+  and the checksum is stable if directory has not changed;
+
+- [x] [collect] Dumped cover filepath is incorrect #4;
+- [x] [collect] Directory checksum should not use computed cover filename with UUID #5;
 - [ ] [rename] Documentation does not describe the "--commit" option #3;
 - [ ] [rename] "rename" command description is wrong #2;